+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+### New Features
+
+#### exonum-node
+
+- Functionality of the `proposer` module was extended. Now, it can also be used
+  for removing transactions from the pool of unconfirmed transactions.
+  The module was renamed to `pool` and related names were updated accordingly.
+  (#1840)
+
 ### Internal Improvements
 
 #### exonum
@@ -72,7 +81,7 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 - Exonum nodes can now customize how they create block proposals. This can be
   used to whitelist / blacklist transaction authors or services, prioritize
   transactions by advanced criteria, implement complex rate limiting, etc.
-  The functionality is available via `proposer` module. (#1820)
+  The functionality is available via `pool` module. (#1820)
 
 - Exonum nodes can now skip block generation at a certain epoch of the consensus
   algorithm. This can be used to keep a "heartbeat" when the network load is low